Alle floats weer omgezet naar doubles. Geheugen verschil marginaal

+++ tactiletriana/src/main/java/nl/utwente/ewi/caes/tactiletriana/simulation/Simulation.java
@@ -95,11 +95,11 @@ public class Simulation extends LoggingEntityBase {
 
     // PROPERTIES
     
-    public float getRadiance() {
+    public double getRadiance() {
         return WeatherData.getInstance().getRadianceProfile()[toTimeStep(getCurrentTime())];
     }
     
-    public float getTemperature() {
+    public double getTemperature() {
         return WeatherData.getInstance().getTemperatureProfile()[toTimeStep(getCurrentTime())];
     }
